requirements.txt
Flask==3.0.3
pywebview==5.1
